+++ src/pages/CreatePaper/CreatePaper.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import SEOMetatag from '../../SEOMetaTag';
+import MetaTag from '../../MetaTag';
 import FormArea from '../../components/CreatePaper/FormArea';
 
 const Wrap = styled.div`
@@ -12,7 +12,7 @@ const Wrap = styled.div`
 export default function CreatePaper() {
   return (
     <Wrap>
-      <SEOMetatag
+      <MetaTag
         title="Rolling - 내 롤링 페이퍼 만들기"
         description="추억의 롤링 페이퍼를 웹에서도 즐길 수 있는 플랫폼 'Rolling'"
       />
